A functioning read/write wiki with to the left the names of the pads
A pad is created by typing a link to a non-existing page in the url-bar:

http://localhost:3000/w/name_of_new_pad

a read-only version of the pad is then available at:

http://localhost:3000/r/name_of_new_pad

Ethertoff.js is based on this django/etherpad application,
by OSP Open Source Publishing:

http://osp.kitchen/tools/ethertoff/

Ethertoff is a simple collaborative web platform, much resembling a wiki
but featuring realtime editing (à la Etherpad). Its output is
constructed with equal love for print and web.

This is the first step for creating a node.js based version,
using the derby.js framework. This promises easier set-up and a smaller
technological foot-print.

This first step possible through the help of these resources:

-  https://github.com/derbyjs/derby-examples (especially the code-mirror
   example)
-  https://github.com/dmapper/derby-tutorials
-  https://github.com/derbyparty/generator-derby